+++ test/unit/workers/author_notifier_worker_test.rb
@@ -1,0 +1,16 @@
+require "test_helper"
+
+class AuthorNotifierWorkerTest < ActiveSupport::TestCase
+  setup do
+    @user = create(:user)
+    @edition = create(:edition)
+  end
+
+  test "calls the AuthorNotifierService" do
+    AuthorNotifierService
+      .expects(:new)
+      .with(@edition, @user)
+
+    AuthorNotifierWorker.new.perform(@edition.id, @user.id)
+  end
+end
